Fix exception in _ce_sender thread
Since the __init__ method of the StateModelsCapability class triggers
a call to trigger_collection_events before the
CollectionEventCapability class is fully initialized, an exception is
raised. This exception can be seen at the end of the pytest report
where it shows the uncaught exceptions raised in started threads.

In order to avoid this exception the order of the Mixin classes is
changed.